+++ NEWS.md
@@ -11,6 +11,8 @@ Explicit multiplication still works.
 * Only variable names can have subscripts. 
 * `latex2fun()` utils were updated so now, for example, any letter can be a function argument. 
 In general the process of transforming a latex expression to a function is more robust.
+* `pi` is not detected as function argument anymore.
+* `latex2fun()` raises an error if there is an assignment within the expression.
 
 ## Deprecation